Validate sample cap in main
Keep rollout execution validation on the main entrypoint so invalid sample caps fail before trainer startup.
